+++ common/src/main/java/net/blay09/mods/waystones/mixin/JigsawPlacementPlacerMixin.java
@@ -31,7 +31,7 @@ public class JigsawPlacementPlacerMixin {
         return WaystonesConfig.getActive().worldGen.spawnInVillages == WaystonesConfigData.VillageWaystoneGeneration.FREQUENT;
     }
 
-    @ModifyArg(method = "tryPlacingChildren(Lnet/minecraft/world/level/levelgen/structure/PoolElementStructurePiece;Lorg/apache/commons/lang3/mutable/MutableObject;IZLnet/minecraft/world/level/LevelHeightAccessor;Lnet/minecraft/world/level/levelgen/RandomState;Lnet/minecraft/world/level/levelgen/structure/pools/alias/PoolAliasLookup;Lnet/minecraft/world/level/levelgen/structure/templatesystem/LiquidSettings;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/core/Registry;getHolder(Lnet/minecraft/resources/ResourceKey;)Ljava/util/Optional;"))
+    @ModifyArg(method = "tryPlacingChildren(Lnet/minecraft/world/level/levelgen/structure/PoolElementStructurePiece;Lorg/apache/commons/lang3/mutable/MutableObject;IZLnet/minecraft/world/level/LevelHeightAccessor;Lnet/minecraft/world/level/levelgen/RandomState;Lnet/minecraft/world/level/levelgen/structure/pools/alias/PoolAliasLookup;Lnet/minecraft/world/level/levelgen/structure/templatesystem/LiquidSettings;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/core/Registry;get(Lnet/minecraft/resources/ResourceKey;)Ljava/util/Optional;"))
     private ResourceKey<StructureTemplatePool> forceWaystonePool(ResourceKey<StructureTemplatePool> resourceKey) {
         if (!waystones$shouldForceWaystone() || waystones$hasWaystone) {
             return resourceKey;
@@ -41,7 +41,7 @@ public class JigsawPlacementPlacerMixin {
         if (poolPath.endsWith("/houses")) {
             ResourceLocation waystonePoolName = ResourceLocation.fromNamespaceAndPath(Waystones.MOD_ID, poolPath.replace("/houses", "/waystones"));
             ResourceKey<StructureTemplatePool> waystonePoolKey = ResourceKey.create(Registries.TEMPLATE_POOL, waystonePoolName);
-            if (pools.getHolder(waystonePoolKey).isPresent()) {
+            if (pools.get(waystonePoolKey).isPresent()) {
                 waystones$hasWaystone = true;
                 return waystonePoolKey;
             }