+++ common/src/main/java/net/blay09/mods/waystones/worldgen/ModWorldGen.java
@@ -9,7 +9,9 @@ import net.blay09.mods.balm.api.event.server.ServerStartedEvent;
 import net.blay09.mods.balm.api.world.BalmWorldGen;
 import net.blay09.mods.balm.api.world.BiomePredicate;
 import net.blay09.mods.waystones.Waystones;
+import net.blay09.mods.waystones.api.WaystoneOrigin;
 import net.blay09.mods.waystones.block.ModBlocks;
+import net.blay09.mods.waystones.block.WaystoneBlock;
 import net.blay09.mods.waystones.config.WaystonesConfig;
 import net.blay09.mods.waystones.config.WaystonesConfigData;
 import net.blay09.mods.waystones.config.WorldGenStyle;
@@ -20,9 +22,10 @@ import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.BiomeTags;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.ai.village.poi.PoiType;
 import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.GenerationStep;
 import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
 import net.minecraft.world.level.levelgen.placement.PlacementModifierType;
@@ -33,6 +36,8 @@ import net.minecraft.world.level.levelgen.structure.templatesystem.StructureProc
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 public class ModWorldGen {
     private static final ResourceLocation waystone = ResourceLocation.fromNamespaceAndPath("waystones", "waystone");
@@ -59,43 +64,43 @@ public class ModWorldGen {
 
         waystonePlacement = worldGen.registerPlacementModifier(id("waystone"), () -> () -> WaystonePlacement.CODEC);
 
-        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.IS_DESERT), GenerationStep.Decoration.VEGETAL_DECORATION, getWaystoneFeature(WorldGenStyle.SANDY));
-        worldGen.addFeatureToBiomes(matchesTag(BiomeTags.IS_JUNGLE), GenerationStep.Decoration.VEGETAL_DECORATION, getWaystoneFeature(WorldGenStyle.MOSSY));
-        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.IS_SWAMP), GenerationStep.Decoration.VEGETAL_DECORATION, getWaystoneFeature(WorldGenStyle.MOSSY));
-        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.IS_MUSHROOM),
+        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.HAS_STRUCTURE_SANDY_WAYSTONE),
+                GenerationStep.Decoration.VEGETAL_DECORATION,
+                getWaystoneFeature(WorldGenStyle.SANDY));
+        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.HAS_STRUCTURE_MOSSY_WAYSTONE),
                 GenerationStep.Decoration.VEGETAL_DECORATION,
                 getWaystoneFeature(WorldGenStyle.MOSSY));
-        worldGen.addFeatureToBiomes(matchesTag(BiomeTags.IS_NETHER),
+        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.HAS_STRUCTURE_BLACKSTONE_WAYSTONE),
                 GenerationStep.Decoration.VEGETAL_DECORATION,
                 getWaystoneFeature(WorldGenStyle.BLACKSTONE));
-        worldGen.addFeatureToBiomes(matchesTag(BiomeTags.IS_END), GenerationStep.Decoration.VEGETAL_DECORATION, getWaystoneFeature(WorldGenStyle.END_STONE));
-        worldGen.addFeatureToBiomes(matchesNeitherTag(List.of(ModBiomeTags.IS_SWAMP,
-                        ModBiomeTags.IS_DESERT,
-                        BiomeTags.IS_JUNGLE,
-                        ModBiomeTags.IS_MUSHROOM,
-                        BiomeTags.IS_NETHER,
-                        BiomeTags.IS_END)),
+        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.HAS_STRUCTURE_END_STONE_WAYSTONE),
+                GenerationStep.Decoration.VEGETAL_DECORATION,
+                getWaystoneFeature(WorldGenStyle.END_STONE));
+        worldGen.addFeatureToBiomes(matchesTag(ModBiomeTags.HAS_STRUCTURE_WAYSTONE),
                 GenerationStep.Decoration.VEGETAL_DECORATION,
                 getWaystoneFeature(WorldGenStyle.DEFAULT));
+
+        worldGen.registerPoiType(id("wild_waystone"), () -> new PoiType(gatherWaystonesOfOrigin(WaystoneOrigin.WILDERNESS), 1, 1));
+        worldGen.registerPoiType(id("village_waystone"), () -> new PoiType(gatherWaystonesOfOrigin(WaystoneOrigin.VILLAGE), 1, 1));
 
         Balm.getEvents().onEvent(ServerStartedEvent.class, event -> setupDynamicRegistries(event.getServer().registryAccess()));
         Balm.getEvents().onEvent(ServerReloadedEvent.class, event -> setupDynamicRegistries(event.getServer().registryAccess()));
     }
 
-    private static BiomePredicate matchesTag(TagKey<Biome> tag) {
-        return (resourceLocation, biome) -> biome.is(tag);
+    private static Set<BlockState> gatherWaystonesOfOrigin(WaystoneOrigin origin) {
+        final var sourceBlocks = List.of(ModBlocks.waystone,
+                ModBlocks.sandyWaystone,
+                ModBlocks.mossyWaystone,
+                ModBlocks.blackstoneWaystone,
+                ModBlocks.endStoneWaystone);
+        return sourceBlocks.stream()
+                .flatMap(it -> it.getStateDefinition().getPossibleStates().stream())
+                .filter(it -> it.getValue(WaystoneBlock.ORIGIN) == origin)
+                .collect(Collectors.toSet());
     }
 
-    private static BiomePredicate matchesNeitherTag(List<TagKey<Biome>> tags) {
-        return (resourceLocation, biome) -> {
-            for (TagKey<Biome> tag : tags) {
-                if (biome.is(tag)) {
-                    return false;
-                }
-            }
-
-            return true;
-        };
+    private static BiomePredicate matchesTag(TagKey<Biome> tag) {
+        return (resourceLocation, biome) -> biome.is(tag);
     }
 
     private static ResourceLocation id(String name) {
@@ -134,15 +139,16 @@ public class ModWorldGen {
     }
 
     private static void addWaystoneStructureToVillageConfig(RegistryAccess registryAccess, String villagePiece, ResourceLocation waystoneStructure, int weight) {
-
-        Holder<StructureProcessorList> emptyProcessorList = registryAccess.registryOrThrow(Registries.PROCESSOR_LIST)
-                .getHolderOrThrow(EMPTY_PROCESSOR_LIST_KEY);
+        Holder<StructureProcessorList> emptyProcessorList = registryAccess.lookupOrThrow(Registries.PROCESSOR_LIST)
+                .getOrThrow(EMPTY_PROCESSOR_LIST_KEY);
         LegacySinglePoolElement piece = StructurePoolElement.legacy(waystoneStructure.toString(), emptyProcessorList)
                 .apply(StructureTemplatePool.Projection.RIGID);
         if (piece instanceof WaystoneStructurePoolElement element) {
             element.waystones$setIsWaystone(true);
         }
-        StructureTemplatePool pool = registryAccess.registryOrThrow(Registries.TEMPLATE_POOL).getOptional(ResourceLocation.withDefaultNamespace(villagePiece)).orElse(null);
+        StructureTemplatePool pool = registryAccess.lookupOrThrow(Registries.TEMPLATE_POOL)
+                .getOptional(ResourceLocation.withDefaultNamespace(villagePiece))
+                .orElse(null);
         if (pool != null) {
             var poolAccessor = (StructureTemplatePoolAccessor) pool;
             // pretty sure this can be an immutable list (when datapacked) so gotta make a copy to be safe.